refactor: extract board state and panels
